requirements.txt
# Core dependencies
mcp>=1.9.0
google-generativeai>=0.8.5
python-dotenv>=1.0.0
uv>=0.1.0

# Development dependencies
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0
